guessing game in python
In this example, the random.randint(1, 100) function is used to generate a random number between 1 and 100. The player is prompted to guess the number and the input is cast to an integer using the int() function. The player's guess is then compared to the randomly generated number, and the player is prompted to guess again if the guess is incorrect. If the player guesses the right number, the program will print a message and exit the loop using the break statement.

You can modify this basic program according to your needs. For example, you can add a limit to the number of tries, or you can add different levels of difficulty by changing the range of the random number.